split tasks/queries.py to execution.py and maintenance.py
