+++ redash/tasks/__init__.py
@@ -1,4 +1,5 @@
 from .general import record_event, version_check, send_mail, sync_user_details
-from .queries import QueryTask, refresh_queries, refresh_schemas, cleanup_query_results, execute_query, empty_schedules
+from .queries import (QueryTask, enqueue_query, execute_query, refresh_queries,
+                      refresh_schemas, cleanup_query_results, empty_schedules)
 from .alerts import check_alerts_for_query
 from .failure_report import send_aggregated_errors
