+++ redash/tasks/queries/__init__.py
@@ -1,0 +1,2 @@
+from .maintenance import refresh_queries, refresh_schemas, cleanup_query_results, empty_schedules
+from .execution import QueryTask, execute_query, enqueue_query
