+++ tests/tasks/test_refresh_queries.py
@@ -1,8 +1,9 @@
 from mock import patch, call, ANY
 from tests import BaseTestCase
-from redash.tasks import refresh_queries
+from redash.tasks.queries.maintenance import refresh_queries
 from redash.models import Query
 
+ENQUEUE_QUERY = 'redash.tasks.queries.maintenance.enqueue_query'
 
 class TestRefreshQuery(BaseTestCase):
     def test_enqueues_outdated_queries(self):
@@ -15,7 +16,7 @@ class TestRefreshQuery(BaseTestCase):
             query_text="select 42;",
             data_source=self.factory.create_data_source())
         oq = staticmethod(lambda: [query1, query2])
-        with patch('redash.tasks.queries.enqueue_query') as add_job_mock, \
+        with patch(ENQUEUE_QUERY) as add_job_mock, \
                 patch.object(Query, 'outdated_queries', oq):
             refresh_queries()
             self.assertEqual(add_job_mock.call_count, 2)
@@ -34,13 +35,13 @@ class TestRefreshQuery(BaseTestCase):
         oq = staticmethod(lambda: [query])
         query.data_source.pause()
         with patch.object(Query, 'outdated_queries', oq):
-            with patch('redash.tasks.queries.enqueue_query') as add_job_mock:
+            with patch(ENQUEUE_QUERY) as add_job_mock:
                 refresh_queries()
                 add_job_mock.assert_not_called()
 
             query.data_source.resume()
 
-            with patch('redash.tasks.queries.enqueue_query') as add_job_mock:
+            with patch(ENQUEUE_QUERY) as add_job_mock:
                 refresh_queries()
                 add_job_mock.assert_called_with(
                     query.query_text, query.data_source, query.user_id,
@@ -59,7 +60,7 @@ class TestRefreshQuery(BaseTestCase):
                 "value": "42",
                 "title": "n"}]})
         oq = staticmethod(lambda: [query])
-        with patch('redash.tasks.queries.enqueue_query') as add_job_mock, \
+        with patch(ENQUEUE_QUERY) as add_job_mock, \
                 patch.object(Query, 'outdated_queries', oq):
             refresh_queries()
             add_job_mock.assert_called_with(
@@ -79,7 +80,7 @@ class TestRefreshQuery(BaseTestCase):
                 "value": 42, # <-- should be text!
                 "title": "n"}]})
         oq = staticmethod(lambda: [query])
-        with patch('redash.tasks.queries.enqueue_query') as add_job_mock, \
+        with patch(ENQUEUE_QUERY) as add_job_mock, \
                 patch.object(Query, 'outdated_queries', oq):
             refresh_queries()
             add_job_mock.assert_not_called()
@@ -100,7 +101,7 @@ class TestRefreshQuery(BaseTestCase):
         dropdown_query = self.factory.create_query(id=100, data_source=None)
 
         oq = staticmethod(lambda: [query])
-        with patch('redash.tasks.queries.enqueue_query') as add_job_mock, \
+        with patch(ENQUEUE_QUERY) as add_job_mock, \
                 patch.object(Query, 'outdated_queries', oq):
             refresh_queries()
             add_job_mock.assert_not_called()
